+++ web/index.js
@@ -12,7 +12,7 @@ import {config, extend} from '@bedrock/web';
 import {createRouter, createWebHistory} from 'vue-router';
 import App from './App.vue';
 import AuditPresentation from './components/AuditPresentation.vue';
-import ButtonView from './components/ButtonView.vue';
+import CHAPIView from './components/CHAPIView.vue';
 import {createHead} from 'unhead';
 import {createI18n} from 'vue-i18n';
 import ErrorView from './components/ErrorView.vue';
@@ -21,7 +21,7 @@ import JsonNode from './components/JsonNode.vue';
 import JsonView from './components/JsonView.vue';
 import LoginView from './components/LoginView.vue';
 import {Notify} from 'quasar';
-import QRView from './components/QRView.vue';
+import OID4VPView from './components/OID4VPView.vue';
 import TestPage from './components/TestPage.vue';
 import TranslateIcon from './components/TranslateIcon.vue';
 import VerificationView from './components/VerificationView.vue';
@@ -37,8 +37,8 @@ brVue.initialize({
 
     app.component('LoginView', LoginView);
     app.component('AuditPresentation', AuditPresentation);
-    app.component('QRView', QRView);
-    app.component('ButtonView', ButtonView);
+    app.component('OID4VPView', OID4VPView);
+    app.component('CHAPIView', CHAPIView);
     app.component('ErrorView', ErrorView);
     app.component('JsonNode', JsonNode);
     app.component('JsonView', JsonView);
@@ -98,6 +98,7 @@ brVue.initialize({
     app.use(VueCookies);
 
     await brQuasar.initialize({app, quasarOptions: {
+      runMode: 'web-client',
       plugins: {
         Notify
       }